rename timings file and snake_case a helper
