+++ src/lib/util/timings.test.ts
@@ -1,7 +1,7 @@
 import {suite} from 'uvu';
 import * as t from 'uvu/assert';
 
-import {create_stopwatch, Timings} from './time.js';
+import {create_stopwatch, Timings} from './timings.js';
 
 /* test_create_stopwatch */
 const test_create_stopwatch = suite('create_stopwatch');
